# openapi.yaml
openapi: 3.0.0
info:
  title: Shell Commands Plugin API
  version: 1.0.0
servers:
  - url: http://localhost:3333
paths:
  /execute/{command}:
    post:
      summary: Executes a shell command on the device
      operationId: executeShellCommand
      parameters:
        - in: path
          name: command
          description: The shell command to execute
          required: true
          schema:
            type: string
      requestBody:
        description: Additional request parameters for the shell command
        content:
          application/json:
            schema:
              type: object
              properties:
                param:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
  /logo.png:
    get:
      summary: Returns the logo of the plugin
      operationId: getPluginLogo
      responses:
        '200':
          description: OK
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Not Found
  /.well-known/ai-plugin.json:
    get:
      summary: Returns the plugin manifest
      operationId: getPluginManifest
      responses:
        '200':
          description: OK
          content:
            text/json:
              schema:
                type: object
        '404':
          description: Not Found
  /openapi.yaml:
    get:
      summary: Returns the OpenAPI specification for the plugin API
      operationId: getPluginOpenAPI
      responses:
        '200':
          description: OK
          content:
            text/yaml:
              schema:
                type: object
        '404':
          description: Not Found